requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
astroid==2.3.3
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
Cython==0.29.14
Django==3.0.2
dlib==19.8.1
dnspython==1.16.0
docutils==0.16
face-recognition==1.2.3
face-recognition-models==0.3.0
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.7
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
idna==2.8
imutils==0.5.3
instaloader==4.2.10
intel-openmp==2019.0
isort==4.3.21
jedi==0.16.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
Kivy==1.11.1
kivy-deps.angle==0.2.0
kivy-deps.glew==0.2.0
kivy-deps.gstreamer==0.2.0
kivy-deps.sdl2==0.2.0
Kivy-Garden==0.1.4
kivymd==0.103.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.1.2
mccabe==0.6.1
mkl==2019.0
mtcnn==0.1.0
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==1.0.1
parso==0.6.1
pathtools==0.1.2
Pillow==7.0.0
playsound==1.2.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pygame==1.9.6
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pypiwin32==223
PyQt4==4.11.4
PyQt5==5.13.0
PyQt5-sip==12.7.1
pyqt5-tools==5.13.0.1.5
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
pywin32==227
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
smmap2==2.0.5
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.8
watchdog==0.10.2
Werkzeug==0.16.0
wincertstore==0.2
winspeech==1.0.1
wrapt==1.11.2
